+++ packages/packages/core/editor-variables/src/components/__tests__/color-variable-edit.test.tsx
@@ -111,61 +111,4 @@ describe( 'ColorVariableEdit', () => {
 			} );
 		} );
 	} );
-
-	// TODO: Pay attention to this test. There is a potential to be flaky.
-	it( '~ should show field-level error when server returns duplicated_label error', async () => {
-		// Arrange.
-		const setValue = jest.fn();
-		const props = {
-			setValue,
-			value: {
-				$$type: colorVariablePropTypeUtil.key,
-				value: 'e-gv-123',
-			},
-			bind: 'color',
-			propType,
-		};
-
-		// Mock API response with duplicated label error
-		const apiErrorResponse = {
-			response: {
-				data: {
-					code: 'duplicated_label',
-				},
-			},
-		};
-
-		( usePropVariablesModule.updateVariable as jest.Mock ).mockRejectedValue( apiErrorResponse );
-
-		// Act.
-		renderControl(
-			<ColorVariableEdit editId="test-variable-id" onClose={ jest.fn() } onSubmit={ jest.fn() } />,
-			props
-		);
-
-		// Change the label to enable the save button
-		const labelField = screen.getByRole( 'textbox', { name: /name/i } );
-		fireEvent.change( labelField, { target: { value: 'new-label' } } );
-
-		// Verify save button is enabled initially
-		const saveButton = screen.getByRole( 'button', { name: /save/i } );
-		expect( saveButton ).toBeEnabled();
-
-		// Trigger the API call by clicking save
-		fireEvent.click( saveButton );
-
-		// Verify that the error response is handled correctly:
-		// 1. Error message is displayed
-		await waitFor( () => {
-			expect(
-				screen.getByText( 'This variable name already exists. Please choose a unique name.' )
-			).toBeInTheDocument();
-		} );
-
-		// 2. Input field has error highlight
-		expect( labelField ).toHaveAttribute( 'aria-invalid', 'true' );
-
-		// 3. Save button is disabled due to error
-		expect( saveButton ).toBeDisabled();
-	} );
 } );
